Made JsonReader an interface to allow composition
