+++ moshi/src/test/java/com/squareup/moshi/JsonAdapterTest.java
@@ -114,7 +114,7 @@ public final class JsonAdapterTest {
       fail();
     } catch (JsonDataException expected) {
       assertThat(expected).hasMessage("Unexpected null at $[1]");
-      assertThat(reader.nextNull()).isNull();
+      assertThat((Object) reader.nextNull()).isNull();
     }
     assertThat(toUpperCase.fromJson(reader)).isEqualTo("C");
     reader.endArray();
