+++ pom.xml
@@ -28,7 +28,7 @@
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <java.version>1.7</java.version>
+    <java.version>1.8</java.version>
 
     <!-- Dependencies -->
     <auto-service.version>1.0-rc5</auto-service.version>
